Refactor board view
